Use raw string to fix Python 3.12 syntax warning
When running under Python v3.12, the `re.match` invocation in
`simpletransform.py` causes a syntax warning, because the regexp string
contains an escape sequence that Python tries to interpret as part of
the string literal:

```bash
python -c 'import re; print(re.match("(translate|scale|rotate|skewX|skewY|matrix)\s*\(([^)]*)\)\s*,?", "translate(3.1416)"))'
<string>:1: SyntaxWarning: invalid escape sequence '\s'
<re.Match object; span=(0, 17), match='translate(3.1416)'>
```

The `\s` is really meant to be taken literally and to be passed to the
regexp engine as is. Use a raw string to achieve that.